Add rocket problem implementation
